@@ -1,7 +1,10 @@
-import React, {useReducer, useCallback, useState} from 'react';
-import {Redirect} from 'react-router-dom';
+import React, { useReducer, useCallback, useState } from 'react';
+import { useSelector } from 'react-redux';
+import { Navigate } from 'react-router-dom';
 import Search from './search/search';
 import Risks from './risks/risks';
+import Backdrop from '../ui/backdrop/backdrop';
+import Spinner from '../ui/spinner/spinner';
 
 const riskReducer = (currentRisk, action) => {
 
@@ -17,10 +20,13 @@ const riskReducer = (currentRisk, action) => {
     }
 };
 
-const risk = React.memo((props) => {
+const risk = React.memo(props => {
 
     const [ riskItems, dispatchRisk ] = useReducer (riskReducer, []);
-    const [redirect, setRedirect] = useState(null);
+    const [ redirect, setRedirect ] = useState(null);
+
+    const loading = useSelector(state => state.risk.loading);
+    const error = useSelector(state => state.risk.error);
 
 
     const filteredRiskHandler = useCallback(filteredRisk => {
@@ -31,30 +37,34 @@ const risk = React.memo((props) => {
     }, [dispatchRisk]);
     
     const onAddRisk = () => {
-        setRedirect(<Redirect to="/registerRisk" />)
+        setRedirect(<Navigate to="/registerRisk" />)
     }
+
+    let spinner = null;
+    if(loading)
+        spinner = <Spinner />;
     
 
     return (
 
-        <section className="section-risk">
+        <section>
             {redirect}
-            <div className="row">
-                <div className="u-margin-bottom-small">
-                    <h2 className="heading-secondary">
-                        Risk
-                    </h2>
+            <Backdrop show={loading} />
+            {spinner}
+            {error &&
+                <div className='alert alert-danger' role='alert'>
+                    {error}
                 </div>
+            }
 
-                <div className="u-margin-bottom-small">
-                    <Search loadRisk={filteredRiskHandler}/>
-                </div>
-
-                <div>
-                    <Risks riskItems={riskItems} onAddRisk={onAddRisk}/>
-                </div>
-
+            <div className='u-margin-bottom-small'>
+                <Search loadRisk={filteredRiskHandler}/>
             </div>
+
+            <div>
+                <Risks riskItems={riskItems} onAddRisk={onAddRisk}/>
+            </div>
+
         </section>
     )
 
