@@ -1,19 +1,18 @@
 import React from 'react';
-import {NavLink} from 'react-router-dom';
+import { NavLink } from 'react-router-dom';
 
 const navigationItem = (props) => (
-	
-	<li className="navigation__item">
-		<NavLink
-			className="navigation__link"
+
+	<li>
+		<NavLink 
 			to={props.link}
-			exact
-			activeClassName=""
-			onClick={props.toggleShow}>
-			{props.children}
+			className={ ({isActive}) => isActive 
+				? 'nav-link text-secondary' 
+				: 'nav-link text-white' }>
+			<i className={ props.icon + ' fs-3 d-block text-sm-center' } />
+			{ props.children }
 		</NavLink>
 	</li>
-
 );
 
 export default navigationItem;
