@@ -2,7 +2,7 @@ import React from 'react';
 
 const backdrop = (props) => (
     props.show
-        ? <div className="backdrop" onClick={props.clicked}></div>
+        ? <div className="modal-backdrop" onClick={props.clicked}></div>
         : null
 );
 
