@@ -2,51 +2,64 @@ import React from 'react';
 
 const features = () => (
 
-    <section className="section-features">
-        
-        <div className="row">
-            <div className="col-one-quarter">
-                <div className="feature-box">
-                    <i className="feature-box__icon icon-basic-world"></i>
-                    <h3 className="heading-tertiary u-margin-bottom-small">Explore risk</h3>
-                    <p>
-                        Praesent ut odio velit. Ut at euismod arcu. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Quisque posuere dui a cursus commodo.
-                    </p>
+    <section className="container px-4 py-5" id="icon-grid">
+        <h2 className="pb-2 border-bottom">Features</h2>
+
+        <div className="row row-cols-1 row-cols-sm-2 row-cols-md-3 row-cols-lg-4 g-4 py-5">
+            <div className="col d-flex align-items-start">
+                <i className='bi-lightbulb fs-1 d-block text-sm-center text-muted flex-shrink-0 me-3' />
+                <div>
+                    <h4 className="fw-bold mb-0">Understanding risk</h4>
+                    <p>Paragraph of text beneath the heading to explain the heading.</p>
                 </div>
             </div>
-
-        
-            <div className="col-one-quarter">
-                <div className="feature-box">
-                    <i className="feature-box__icon icon-basic-upload"></i>
-                    <h3 className="heading-tertiary u-margin-bottom-small">Register a risk</h3>
-                    <p>
-                        Praesent ut odio velit. Ut at euismod arcu. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Quisque posuere dui a cursus commodo.
-                    </p>
+            <div className="col d-flex align-items-start">
+                <i className='bi-graph-down-arrow fs-1 d-block text-sm-center text-muted flex-shrink-0 me-3' />
+                <div>
+                    <h4 className="fw-bold mb-0">Risk profile</h4>
+                    <p>Paragraph of text beneath the heading to explain the heading.</p>
                 </div>
             </div>
-        
-
-        
-            <div className="col-one-quarter">
-                <div className="feature-box">
-                    <i className="feature-box__icon icon-basic-calculator"></i>
-                    <h3 className="heading-tertiary u-margin-bottom-small">Review risk</h3>
-                    <p>
-                        Praesent ut odio velit. Ut at euismod arcu. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Quisque posuere dui a cursus commodo.
-                    </p>
+            <div className="col d-flex align-items-start">
+                <i className='bi-percent fs-1 d-block text-sm-center text-muted flex-shrink-0 me-3' />
+                <div>
+                    <h4 className="fw-bold mb-0">Risk coverage</h4>
+                    <p>Paragraph of text beneath the heading to explain the heading.</p>
                 </div>
             </div>
-        
-
-        
-            <div className="col-one-quarter">
-                <div className="feature-box">
-                    <i className="feature-box__icon icon-basic-folder-multiple"></i>
-                    <h3 className="heading-tertiary u-margin-bottom-small">Reports</h3>
-                    <p>
-                        Praesent ut odio velit. Ut at euismod arcu. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Quisque posuere dui a cursus commodo.
-                    </p>
+            <div className="col d-flex align-items-start">
+                <i className='bi-sun fs-1 d-block text-sm-center text-muted flex-shrink-0 me-3' />
+                <div>
+                    <h4 className="fw-bold mb-0">Opportunities</h4>
+                    <p>Paragraph of text beneath the heading to explain the heading.</p>
+                </div>
+            </div>
+            <div className="col d-flex align-items-start">
+                <i className='bi-ui-checks-grid fs-1 d-block text-sm-center text-muted flex-shrink-0 me-3' />
+                <div>
+                    <h4 className="fw-bold mb-0">Incident management</h4>
+                    <p>Paragraph of text beneath the heading to explain the heading.</p>
+                </div>
+            </div>
+            <div className="col d-flex align-items-start">
+                <i className='bi-balloon-heart fs-1 d-block text-sm-center text-muted flex-shrink-0 me-3' />
+                <div>
+                    <h4 className="fw-bold mb-0">Risk engagement</h4>
+                    <p>Paragraph of text beneath the heading to explain the heading.</p>
+                </div>
+            </div>
+            <div className="col d-flex align-items-start">
+                <i className='bi-file-earmark-person fs-1 d-block text-sm-center text-muted flex-shrink-0 me-3' />
+                <div>
+                    <h4 className="fw-bold mb-0">Individual profiles</h4>
+                    <p>Paragraph of text beneath the heading to explain the heading.</p>
+                </div>
+            </div>
+            <div className="col d-flex align-items-start">
+                <i className='bi-bar-chart-line fs-1 d-block text-sm-center text-muted flex-shrink-0 me-3' />
+                <div>
+                    <h4 className="fw-bold mb-0">Reporting</h4>
+                    <p>Paragraph of text beneath the heading to explain the heading.</p>
                 </div>
             </div>
         </div>
