@@ -1,37 +1,30 @@
 import React from 'react';
-import { updateObject } from '../../../shared/utility';
 
-const range = React.memo(props => {
+const range = (props) => {
 
-    const {config, setConfig} = props;
-    
-    // const [range, setRange] = useState(1);
+    const { config, setConfig } = props;
 
     const changeHandler = (event) => {
-        setConfig(
-            updateObject(config, {
+        setConfig({
+            ...config,
                 value: event.target.value
-            })
-        );
+        });
     };
 
     return (
-        <div className="register-risk__range range">
-            <div className="register-risk__panel-item range">
-            <label className="form__label">{config.label + ' - ' + config[config.value].title}</label>
-                <input 
-                    type="range" 
-                    min={config.min} 
-                    max={config.max}
-                    step={config.step}
-                    value={config.value} onChange={event => changeHandler(event)} />
-                <span>{config[config.value].description}</span>
-            </div>
-            
+        <div className='range text-center'>
+            <input
+                className='mb-5'
+                type="range" 
+                min={config.min} 
+                max={config.max}
+                step={config.step}
+                value={config.value}
+                onChange={event => changeHandler(event)} />
+            <span>{config[config.value].description}</span>
         </div>
     );
 
-    // --style='--min:0; --max:10000; --step:100; --value:200; --text-value:"200"; --prefix:"$"'
-});
+}   ;
 
 export default range;
