@@ -9,9 +9,9 @@ export const AUTH_LOGOUT = 'AUTH_LOGOUT';
 export const AUTH_STATE_RESET = 'AUTH_STATE_RESET';
 export const AUTH_PATCH_AVATAR_URL = 'AUTH_PATCH_AVATAR_URL';
 export const AUTH_DELETE_AVATAR_URL = 'AUTH_DELETE_AVATAR_URL';
-export const AUTH_POST_DISPLAY_NAME = 'AUTH_POST_DISPLAY_NAME';
-export const AUTH_POST_EMAIL = 'AUTH_POST_EMAIL';
-export const AUTH_POST_PASSWORD = 'AUTH_POST_PASSWORD';
+export const AUTH_PATCH_DISPLAY_NAME = 'AUTH_PATCH_DISPLAY_NAME';
+export const AUTH_PATCH_EMAIL = 'AUTH_PATCH_EMAIL';
+export const AUTH_PATCH_PASSWORD = 'AUTH_PATCH_PASSWORD';
 export const AUTH_POST_FORGOTTEN_PASSWORD = 'AUTH_POST_FORGOTTEN_PASSWORD';
 
 // Feedback defined action types
@@ -34,8 +34,11 @@ export const INTELLIVERSE_REDIRECT_PATH = 'INTELLIVERSE_REDIRECT_PATH';
 
 // Risk defined action types
 export const RISK_START = 'RISK_START';
-export const RISK_SUCCESS = 'RISK_SUCCESS';
+export const RISK_GET_SUCCESS = 'RISK_GET_SUCCESS';
+export const RISK_POST_SUCCESS = 'RISK_POST_SUCCESS';
+export const RISK_PATCH_SUCCESS = 'RISK_PATCH_SUCCESS';
 export const RISK_FINISH = 'RISK_FINISH';
+export const RISK_SELECT = 'RISK_SELECT';
 export const RISK_FAIL = 'RISK_FAIL';
 export const RISK_ERROR_RESET = 'RISK_ERROR_RESET';
 export const RISK_STATE_RESET = 'RISK_ERROR_RESET';
